modularize-4: split audit.rs into audit/ directory
Per refactor-plan-2026-05-20.md candidate #4: the 1485-line
crates/octravpn-node/src/audit.rs interleaved a sync write path, an
async batched-fsync flusher with bounded queue + inline fallback, the
offline verifier, the analytics tap, the HMAC chain math, and the
process-lifetime counters. Sibling agents stepped on it repeatedly.

Split into:

  - audit/mod.rs        — pub(crate) AuditLog handle + crate-public
                          re-exports + inline_fallback_total() accessor
  - audit/inner.rs      — owned Inner state + AuditCounters
                          (lock-free atomics, outside the mutex)
  - audit/log.rs        — AuditRecord, ChainedLine envelope, sync
                          open/write, shared write_inner_direct
  - audit/batched.rs    — async flusher, FlusherCmd, open_batched
                          (_with_cap), bounded mpsc, inline fallback,
                          write_async, flush_and_close,
                          record_receipt_signed, queue-knob constants
  - audit/chain.rs      — pure chain_step + load_or_create_key + date
  - audit/verify.rs     — FileVerifyReport, FileVerifyError
                          (#[non_exhaustive]), offline verify_file
  - audit/tap.rs        — analytics tap (with_analytics_tap,
                          tap_publish)
  - audit/test_util.rs  — shared #[cfg(test)] helpers
  - audit/README.md     — threading model + durability ladder +
                          lock-order rules + on-disk format contract

Constraints preserved:

  - On-disk format unchanged (audit-YYYY-MM-DD.jsonl round-trip
    byte-identical).
  - pub(crate) surface unchanged — AuditLog, AuditRecord,
    FileVerifyReport, DEFAULT_BATCH_SIZE, DEFAULT_BATCH_INTERVAL_MS,
    DEFAULT_BATCH_QUEUE_CAP, chain_step all re-exported from
    audit/mod.rs. inline_fallback_total() preserved for /metrics.
  - DEFAULT_BATCH_SIZE=64, DEFAULT_BATCH_INTERVAL_MS=100,
    DEFAULT_BATCH_QUEUE_CAP=8192 unchanged.
  - All 19 inline tests preserved (13 baseline + 6 OOM-3 bounded-queue
    tests from Fix-5) and pass.
  - cargo build -p octravpn-node clean.
  - hub.rs, audit_cli.rs, control/handlers/metrics.rs, cli_ops.rs,
    main.rs all build without source changes.

Cross-module test surfacing the split's value: verify::tests::
receipt_signed_entry_round_trips — exercises batched.rs emit helper
record_receipt_signed + log.rs ChainedLine envelope + this module's
verify_file in one flow. After the split its scope is unambiguous
("the three modules agree on the receipt_signed schema") rather than
buried in a 1485-line file.
